Deliver as many webhooks as possible.
If one fails, this only raises the error at the end, giving others,
potentially more successful ones, a reason to pass.

Motivation is that other webhook deliveries shouldn't suffer from
one webhook in the chain not being able to deliver.

The downside is a possible duplication in deliveries for the ones
that have succeeded.

Fixes travis-ci/travis-ci#2356.