improve with how we deal with errors from GitHub
if we get a 401 error (token auth error), log a line and move on.

if we get a 403 error (rate limit exceeded), only raise an error if in production.

if we get a 404 error (repo not found), log a line and move on as this is most likely an auth error.