Raise exception if token is invalid
Currently, when an invalid token is passed to a webhook task,
if we are not using Basic Auth, we will attempt to craft an
Authorization header that users then compare to ensure that the request
has come from Travis.

We use the `token` as well as the `slug` to calculate the value of the
Authorization header. When token is nil/"", this breaks consuming
clients' code, as the signature we calculate is no longer valid.

To ensure this doesn't happen again in the future, we'll raise an
exception whenever the token we are passed is unusable. This won't
impact other code paths, but will let us know if this problem crops up
again in the future.