Update email validity check
First, note that http://is.gd/Dzd6fp, which is a redirect to
http://my.rails-royce.org/2010/07/21/email-validation-in-ruby-on-rails-without-regexp,
is no longer valid. It is not clear what it said, but based on the
comment, it is safe to assume that it suggested using the private
`#tree` method to access `Mail::Address`'s internal data.

Second, `mail` 2.6.0 switched the address parser to a Ragel-based one,
so this invasive method is no longer valid.

Based on the following data structure from `mail` 2.5.5:

```
address.__send__(:tree).domain.dot_atom_text.elements
=> [SyntaxNode+DotAtomText0 offset=3, "mail." (domain_text):
  SyntaxNode offset=3, "mail":
    SyntaxNode offset=3, "m"
    SyntaxNode offset=4, "a"
    SyntaxNode offset=5, "i"
    SyntaxNode offset=6, "l"
  SyntaxNode offset=7, ".",
 SyntaxNode+DotAtomText0 offset=8, "org" (domain_text):
  SyntaxNode offset=8, "org":
    SyntaxNode offset=8, "o"
    SyntaxNode offset=9, "r"
    SyntaxNode offset=10, "g"
  SyntaxNode offset=11, ""]
```

we deduce that the last criterion is that the "domain" part of the
given email address has more than one parts.