Remove `GH::TokenCheck` from GH call stack
While checking the validity of the token we are going to use is
a good idea, it seems that we are making quite a few of them,
consuming a lot of API calls, often unnecessarily.

It is possible to memoize a token's validity for some time in Redis,
if we want to still check it.